Updates to Bibliography component, Citation component, and .mdx files.
Updated the Bibliography component logic to automatically determine the url pathname for the page that's being viewed and generate the file path to the bibliogrphy file in the static/bibliographies folder, eliminating the need for the user to specify a bibFile prop.

Removed a console log from the Citation component to return the bibFilePath.

Updated all .mdx files to remove unnecessary props from React components.